Development: Move Jenkins CI backend into separate module
Mechanical extraction following the pattern set by notification/calendar/account/course/admin.

Server side:
- Move 16 main classes from programming/service/jenkins/{,build_plan,jobs} to a new top-level
  jenkins/service/{,build_plan,jobs} package. All classes keep their existing @Profile(PROFILE_JENKINS)
  gating; Spring picks them up by package scanning.
- Move the 3 Jenkins integration tests to jenkins/service[/build_plan] and rename them from *Test to
  *IntegrationTest so they match the module test-architecture rule (they already extend
  AbstractProgrammingIntegrationJenkinsLocalVCTest).
- Add the six standard module ArchUnit tests (Service / CodeStyle / Repository / Resource /
  EntityUsage / Test). Jenkins has no REST controllers, repositories, or DTOs of its own — the empty
  rules lock in the zero-violation state.
- Update every external import (RestTemplateConfiguration, JenkinsRequestMockProvider,
  ProgrammingExerciseTestService, AbstractSpringIntegrationJenkinsLocalVCTestBase,
  AbstractProgrammingIntegrationJenkinsLocalVCTest, RepositoryProgrammingExerciseParticipationJenkinsIntegrationTest)
  to the new package.

Build / docs:
- Add per-module Jacoco threshold for jenkins (conservative initial baseline; ratchet up later).
- CLAUDE.md: list jenkins as a separate module and clarify programming's scope.
- CODEOWNERS: jenkins module is owned by @ls1intum/programming-maintainers.

Client side: no changes. The Jenkins-touching client code (build-plan editor) is shared with
LocalCI and remains in app/programming/ — documented in the plan.

Plan source: docs/superpowers/plans/2026-05-26-notification-and-programming-module-extractions.md
sub-project B.
