+++ CLAUDE.md
@@ -111,7 +111,8 @@ Organized by feature module:
 - `core/` - Configuration, security base, utilities, base entities
 - `account/` - User, authority, passkey, account REST, authentication, LDAP
 - `exercise/` - Base exercise functionality
-- `programming/` - Programming exercises with CI/CD
+- `programming/` - Programming exercises (lifecycle, grading, repositories)
+- `jenkins/` - Jenkins CI backend connector (used by `programming` exercises configured to use Jenkins)
 - `quiz/` - Quiz exercises
 - `modeling/` - UML diagram exercises
 - `text/` - Text exercises
