+++ gradle/jacoco.gradle
@@ -29,6 +29,9 @@ ext {
         "fileupload"    :          ["INSTRUCTION": 0.924, "CLASS":  1],
         "hyperion"      :          ["INSTRUCTION": 0.646, "CLASS":  9],
         "iris"          :          ["INSTRUCTION": 0.811, "CLASS": 21],
+        // New module extracted from programming in this PR. Conservative initial baseline; ratchet up
+        // once CI measures the actual coverage of the extracted Jenkins connector code.
+        "jenkins"       :          ["INSTRUCTION": 0.750, "CLASS":  5],
         "lecture"       :          ["INSTRUCTION": 0.891, "CLASS":  1],
         "lti"           :          ["INSTRUCTION": 0.860, "CLASS":  2],
         "modeling"      :          ["INSTRUCTION": 0.907, "CLASS":  1],
