+++ src/test/java/de/tum/cit/aet/artemis/jenkins/architecture/JenkinsEntityUsageArchitectureTest.java
@@ -1,0 +1,30 @@
+package de.tum.cit.aet.artemis.jenkins.architecture;
+
+import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleEntityUsageArchitectureTest;
+
+/**
+ * Jenkins is a connector module with no REST controllers or DTOs of its own. The base class still runs
+ * (with allowEmptyShould) to lock in the zero-violation state and catch future regressions.
+ */
+class JenkinsEntityUsageArchitectureTest extends AbstractModuleEntityUsageArchitectureTest {
+
+    @Override
+    public String getModulePackage() {
+        return ARTEMIS_PACKAGE + ".jenkins";
+    }
+
+    @Override
+    protected int getExpectedEntityReturnViolations() {
+        return 0;
+    }
+
+    @Override
+    protected int getExpectedEntityInputViolations() {
+        return 0;
+    }
+
+    @Override
+    protected int getExpectedDtoEntityFieldViolations() {
+        return 0;
+    }
+}
