+++ src/test/java/de/tum/cit/aet/artemis/jenkins/architecture/JenkinsServiceArchitectureTest.java
@@ -1,0 +1,11 @@
+package de.tum.cit.aet.artemis.jenkins.architecture;
+
+import de.tum.cit.aet.artemis.shared.architecture.module.AbstractModuleServiceArchitectureTest;
+
+class JenkinsServiceArchitectureTest extends AbstractModuleServiceArchitectureTest {
+
+    @Override
+    public String getModulePackage() {
+        return ARTEMIS_PACKAGE + ".jenkins";
+    }
+}
